+++ src/aks-preview/HISTORY.rst
@@ -11,10 +11,16 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+* Add add-on support for `ApplicationLoadBalancer` (Application Gateway for Containers)
+- AKS Create and Update parameters for `--enable-application-load-balancer` and `--disable-application-load-balancer`
+ - `az aks applicationloadbalancer enable`: Enable Application Load Balancer add-on for an existing cluster.
+ - `az aks applicationloadbalancer disable`: Disable Application Load Balancer add-on for an existing cluster.
+ - `az aks applicationloadbalancer update`: Update Application Load Balancer add-on for an existing cluster.
 
 19.0.0b18
 +++++++
 * Vendor new SDK and bump API version to 2025-10-02-preview.
+* Add option `--gpu-driver` to `az aks nodepool update` to select skipping GPU driver installation.
 
 19.0.0b17
 +++++++